+++ Source/Core/StringUtils.h
@@ -253,8 +253,8 @@ RIFT_API NO_DISCARD u8   ToUpper(u8 Char)         CONST_FN;
 RIFT_API NO_DISCARD u8   ToLower(u8 Char)         CONST_FN;
 RIFT_API NO_DISCARD u8   ToForwardSlash(u8 Char)  CONST_FN;
 RIFT_API NO_DISCARD u8   ToBackSlash(u8 Char)     CONST_FN;
-RIFT_API NO_DISCARD uchar DigitToHexChar(u8 Val)  CONST_FN;
-RIFT_API NO_DISCARD uchar DigitToHexCharUpper(u8 Val)  CONST_FN;
+RIFT_API NO_DISCARD uchar DigitToHexChar(u32 Val)  CONST_FN;
+RIFT_API NO_DISCARD uchar DigitToHexCharUpper(u32 Val)  CONST_FN;
 
 RIFT_API NO_DISCARD u8 Integer_CountDigits(u64 Value) CONST_FN;
 RIFT_API NO_DISCARD u8 Integer_CountDigits_Signed(i64 Value) CONST_FN;
